refactor: move functionality to base and nwd modules
